requirements.txt
Django==1.7
django-reversion==1.8.5
django-simple-history==1.5.3
